+++ src/Branch/UE3/GIGANTIC/EngineBranchGigantic.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using UELib.Branch.UE3.GIGANTIC.Core.Classes;
 using UELib.Branch.UE3.GIGANTIC.Tokens;
 using UELib.Core.Tokens;
 
@@ -17,9 +18,16 @@ namespace UELib.Branch.UE3.GIGANTIC
         {
             JsonImport = 0x00008000U,
         }
-        
+
         public EngineBranchGigantic(BuildGeneration generation) : base(BuildGeneration.UE3)
         {
+        }
+
+        public override void PostDeserializePackage(UnrealPackage linker, IUnrealStream stream)
+        {
+            base.PostDeserializePackage(linker, stream);
+
+            linker.AddClassType("JsonNodeRoot", typeof(UJsonNodeRoot));
         }
 
         protected override TokenMap BuildTokenMap(UnrealPackage linker)